create listing

+++ src/main/java/org/tiere/dto/Account.java
@@ -1,4 +1,4 @@
 package org.tiere.dto;
 
-public record Account(Long id, String email) {
+public record Account(Integer id, String email) {
 }
